Add CI tests for SFT/PEFT (#10632)
* fix import


* add ci tests


* Apply isort and black reformatting


* fix typo


* fix typo


* reorder and fix path


* fix yaml


* fix failed unit test


* fix failed unit test


* make ckpt format backwards compatible


* Apply isort and black reformatting


* Revert "Apply isort and black reformatting"

This reverts commit 4677597e9c09bbafa025562582dcf414d3f3c80a.

* Revert "make ckpt format backwards compatible"

This reverts commit ba9639fc2fffdb79599d6eca7b02323957bf6106.

* move ckpt loading path change logic one level up


* add comment


* unintended changes


* remove old assert


* fix peft resume with new ch


* lora fix from #10850


* Apply isort and black reformatting


---------

+++ nemo/lightning/io/artifact/file.py
@@ -17,6 +17,10 @@ class PathArtifact(Artifact[Path]):
 
 class FileArtifact(Artifact[str]):
     def dump(self, value: str, absolute_dir: Path, relative_dir: Path) -> str:
+        if not pathize(value).exists():
+            # This is Artifact is just a string.
+            self.skip = True
+            return value
         new_value = copy_file(value, absolute_dir, relative_dir)
         return str(new_value)
